Refactor dataloders into seperate files
